# docs/swagger.yaml
basePath: /
definitions:
...
  properties.CreatePropertyInput:
    properties:
      clauses:
        items:
          $ref: '#/definitions/properties.CreatePropertyClauseInput'
        type: array
      commercial:
        $ref: '#/definitions/properties.CreateCommercialInput'
      description:
        example: Spacious residential property near downtown
        type: string
      is_featured:
        example: false
        type: boolean
      location:
        $ref: '#/definitions/properties.CreateLocationInput'
      lot_area:
        example: 200
        type: number
      modality_id:
        example: 1
        type: integer
      property_type_id:
        example: 1
        type: integer
      rent_prices:
        items:
          $ref: '#/definitions/properties.CreateRentPriceInput'
        type: array
      residential:
        $ref: '#/definitions/properties.CreateResidentialInput'
      sale_price:
        $ref: '#/definitions/properties.CreateSalePriceInput'
      services:
        example:
        - 1
        - 3
        - 7
        items:
          type: integer
        type: array
      subtype:
        example: residential
        type: string
      title:
        example: Casa en Xalapa
        type: string
    type: object
  properties.CreatePropertyResult:
    properties:
      data:
        $ref: '#/definitions/properties.CreatePropertyResultData'
    type: object
  properties.CreatePropertyResultData:
    properties:
      property_uuid:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
...
  properties.DeletePropertyInput:
    properties:
      confirm:
        example: true
        type: boolean
    type: object
  properties.GetPropertyClausesResult:
    properties:
      data:
        items:
          $ref: '#/definitions/properties.PropertyClauseData'
        type: array
    type: object
...
  properties.GetPropertyHistoryResult:
    properties:
      data:
        items:
          $ref: '#/definitions/properties.PropertyStatusHistoryData'
        type: array
    type: object
  properties.GetPropertyPhotosResult:
    properties:
      data:
        items:
          $ref: '#/definitions/properties.PropertyPhotoData'
        type: array
    type: object
...
  properties.GetPropertyPricesHistoryResult:
    properties:
      data:
        items:
          $ref: '#/definitions/properties.PropertyPriceHistoryData'
        type: array
    type: object
  properties.GetPropertyPricesResult:
    properties:
      data:
        $ref: '#/definitions/properties.GetPropertyPricesData'
    type: object
  properties.GetPropertyResult:
    properties:
      data:
        $ref: '#/definitions/properties.GetPropertyData'
    type: object
...
  properties.GetPropertyServicesResult:
    properties:
      data:
        $ref: '#/definitions/properties.GetPropertyServicesData'
    type: object
  properties.ListPropertiesMeta:
    properties:
      current_page:
        type: integer
      has_next:
        type: boolean
      has_prev:
        type: boolean
      page_size:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
  properties.ListPropertiesResult:
    properties:
      data:
        items:
          $ref: '#/definitions/properties.PropertyCardData'
        type: array
      meta:
        $ref: '#/definitions/properties.ListPropertiesMeta'
    type: object
...
  properties.UpdatePropertyClausesInput:
    properties:
      clauses:
        items:
          $ref: '#/definitions/properties.CreatePropertyClauseInput'
        type: array
    type: object
  properties.UpdatePropertyInput:
    properties:
      commercial:
        $ref: '#/definitions/properties.UpdateCommercialInput'
      description:
        type: string
      is_featured:
        type: boolean
      location:
        $ref: '#/definitions/properties.UpdateLocationInput'
      lot_area:
        type: number
      residential:
        $ref: '#/definitions/properties.UpdateResidentialInput'
      title:
        type: string
    type: object
  properties.UpdatePropertyPhotosInput:
    properties:
      photos:
        items:
          $ref: '#/definitions/properties.UpdatePhotoMetadataInput'
        type: array
    type: object
  properties.UpdatePropertyPricesInput:
    properties:
      rent_prices:
        items:
          $ref: '#/definitions/properties.UpdateRentPriceInput'
        type: array
      sale_price:
        $ref: '#/definitions/properties.UpdateSalePriceInput'
    type: object
  properties.UpdatePropertyResult:
    properties:
      message:
        type: string
    type: object
  properties.UpdatePropertyServicesInput:
    properties:
      service_ids:
        items:
          type: integer
        type: array
    type: object
...
  shared.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
  uploads.UploadPhotoResult:
    properties:
      photo_id:
        type: integer
      storage_key:
        type: string
      url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API de Spazio Backend
  title: Spazio API
  version: "1.0"
paths:
...
  /api/v1/properties:
    get:
      description: Returns a paginated list of property cards. Administrators can
        view non-deleted properties, while agents can only view properties assigned
        in property_agents. Supports filtering by search query, status, property type,
        modality, location (country, state, city), price range, and minimum bedrooms.
        Price selection logic prioritizes sale price, then the best current rent price
        (Monthly > Annual > Weekly > Daily).
      parameters:
      - default: 1
        description: Page number (starts at 1)
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: page_size
        type: integer
      - description: Search term (matches title, address, city, state, country)
        in: query
        name: q
        type: string
      - collectionFormat: csv
        description: Filter by status IDs (defaults to Available=2 when omitted)
        in: query
        items:
          type: integer
        name: status_id
        type: array
      - description: Filter by property type ID
        in: query
        name: property_type_id
        type: integer
      - description: Filter by modality ID
        in: query
        name: modality_id
        type: integer
      - description: Filter by country ID
        in: query
        name: country_id
        type: integer
      - description: Filter by state ID
        in: query
        name: state_id
        type: integer
      - description: Filter by city ID
        in: query
        name: city_id
        type: integer
      - description: Minimum price filter
        in: query
        name: min_price
        type: number
      - description: Maximum price filter
        in: query
        name: max_price
        type: number
      - description: Minimum bedrooms filter (Residential only)
        in: query
        name: min_bedrooms
        type: integer
      - description: 'Sort by: created_at, title, price'
        in: query
        name: sort
        type: string
      - description: 'Sort order: asc, desc'
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of property cards
          schema:
            $ref: '#/definitions/properties.ListPropertiesResult'
        "400":
          description: Invalid input parameters
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "401":
          description: Missing or invalid authenticated session
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List properties with advanced filters
      tags:
      - Properties
    post:
      consumes:
      - application/json
      description: Registers a property and all related records in a single database
        transaction. The backend generates the property UUID and stores subtype, location,
        pricing, services, and clauses atomically. The authenticated user is set as
        the owner.
      parameters:
      - description: Property payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/properties.CreatePropertyInput'
      produces:
      - application/json
      responses:
        "201":
          description: Property created
          schema:
            $ref: '#/definitions/properties.CreatePropertyResult'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Register a new property
      tags:
      - Properties
  /api/v1/properties/{uuid}:
    delete:
      consumes:
      - application/json
      description: Soft deletes a property by UUID. Requires confirm=true. The authenticated
        admin is recorded as the author of the status change. The operation deletes
        linked photo objects from storage before applying database updates. Only properties
        with available status can be deleted.
      parameters:
      - description: Property UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Delete payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/properties.DeletePropertyInput'
      produces:
      - application/json
      responses:
        "200":
          description: Property deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete property
      tags:
      - Properties
    get:
      description: Returns property base data, subtype, and location for the given
        UUID. Administrators can view all fields including registered_by. Agents can
        only access assigned properties and do not receive registered_by.
      parameters:
      - description: Property UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Property data
          schema:
            $ref: '#/definitions/properties.GetPropertyResult'
        "400":
          description: Invalid path parameter
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "401":
          description: Missing or invalid authenticated session
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get property by UUID
      tags:
      - Properties
    patch:
      consumes:
      - application/json
      description: |-
        Updates property base data, subtype and location.
        Requires an authenticated admin session.
      parameters:
      - description: Property UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/properties.UpdatePropertyInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/properties.UpdatePropertyResult'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update property
      tags:
      - Properties
...
  /api/v1/properties/{uuid}/clauses:
    get:
      description: Returns the clauses linked to a property by UUID. When the property
        has no clauses, the response contains an empty data array.
      parameters:
      - description: Property UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Property clauses
          schema:
            $ref: '#/definitions/properties.GetPropertyClausesResult'
        "400":
          description: Invalid path parameter
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      summary: List property clauses
      tags:
      - Properties
    put:
      consumes:
      - application/json
      description: Replaces all clauses linked to a property by UUID. When the payload
        omits clauses or sends an empty array, all linked clauses are removed.
      parameters:
      - description: Property UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Property clauses payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/properties.UpdatePropertyClausesInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      summary: Replace property clauses
      tags:
      - Properties
  /api/v1/properties/{uuid}/history:
    get:
      description: Returns the chronological history of status changes for a specific
        property. Requires an authenticated admin session.
      parameters:
      - description: Property UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Chronological status history retrieved successfully
          schema:
            $ref: '#/definitions/properties.GetPropertyHistoryResult'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "401":
          description: Missing or invalid authenticated session
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get property status history
      tags:
      - Properties
  /api/v1/properties/{uuid}/photos:
    get:
      description: Returns the photo metadata linked to a property by UUID ordered
        by sort_order ascending. When the property has no photos, the response contains
        an empty data array.
      parameters:
      - description: Property UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Property photos
          schema:
            $ref: '#/definitions/properties.GetPropertyPhotosResult'
        "400":
          description: Invalid path parameter
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      summary: List property photos
      tags:
      - Properties
    put:
      consumes:
      - application/json
      description: Replaces the stored metadata of the photos linked to a property
        by UUID. Exactly one photo must be marked as cover when the payload contains
        photos. When the payload omits photos or sends an empty array, all linked
        photos are removed.
      parameters:
      - description: Property UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Property photos payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/properties.UpdatePropertyPhotosInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      summary: Replace property photos
      tags:
      - Properties
  /api/v1/properties/{uuid}/prices:
    get:
      description: Returns the active prices of a property by UUID. Sale price and
        rent prices include only currently active records (is_current=true). If no
        active sale price exists, sale_price is null. If no active rent prices exist,
        rent_prices is an empty array.
      parameters:
      - description: Property UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Property prices
          schema:
            $ref: '#/definitions/properties.GetPropertyPricesResult'
        "400":
          description: Invalid path parameter
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      summary: Get property prices
      tags:
      - Properties
    put:
      consumes:
      - application/json
      description: Updates property prices by UUID. Only processes prices in the payload;
        unmodified prices remain unchanged. When the amount of a price changes, a
        new price record is created and the old one is marked as inactive. Amounts
        must be greater than 0. Currency is not editable.
      parameters:
      - description: Property UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Property prices payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/properties.UpdatePropertyPricesInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      summary: Update property prices
      tags:
      - Properties
  /api/v1/properties/{uuid}/prices/history:
    get:
      description: Returns the complete price history for a property. Requires admin
        role.
      parameters:
      - description: Property UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Property prices history
          schema:
            $ref: '#/definitions/properties.GetPropertyPricesHistoryResult'
        "400":
          description: Invalid path parameter
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "401":
          description: Missing or invalid authenticated session
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get property prices history
      tags:
      - Properties
  /api/v1/properties/{uuid}/services:
    get:
      description: Returns the service IDs linked to a property by UUID. When the
        property has no services, the response contains an empty service_ids array.
      parameters:
      - description: Property UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Property services
          schema:
            $ref: '#/definitions/properties.GetPropertyServicesResult'
        "400":
          description: Invalid path parameter
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      summary: List property services
      tags:
      - Properties
    put:
      consumes:
      - application/json
      description: Replaces all services linked to a property by UUID. When the payload
        omits service_ids or sends an empty array, all linked services are removed.
      parameters:
      - description: Property UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Property services payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/properties.UpdatePropertyServicesInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      summary: Replace property services
      tags:
      - Properties
...
  /api/v1/uploads/properties/{property_uuid}/photos:
    post:
      consumes:
      - multipart/form-data
      description: Uploads a property photo to object storage and registers metadata.
        Max size is 5MB and allowed types are image/jpeg, image/png, and image/webp.
      parameters:
      - description: Property UUID
        in: path
        name: property_uuid
        required: true
        type: string
      - description: Image file
        in: formData
        name: file
        required: true
        type: file
      - description: Optional label
        in: formData
        name: label
        type: string
      - description: Optional alt text
        in: formData
        name: alt_text
        type: string
      - default: 0
        description: Sort order
        in: formData
        name: sort_order
        type: integer
      - default: false
        description: Mark as cover
        in: formData
        name: is_cover
        type: boolean
      produces:
      - application/json
      responses:
        "201":
          description: Photo uploaded
          schema:
            $ref: '#/definitions/uploads.UploadPhotoResult'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      summary: Upload property photo
      tags:
      - Uploads
...
swagger: "2.0"